+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Patch notes
 
-Current vue-extended-multiselect version: **0.5.3**
+Current vue-extended-multiselect version: **0.5.4**
 
 ***
 
@@ -116,3 +116,7 @@ Current vue-extended-multiselect version: **0.5.3**
 ### 0.5.3 (2024-04-04)
 
 * Fixing bug with incorrect options list height calculation.
+
+### 0.5.4 (2024-04-04)
+
+* Fixing bug with incorrect options list appearance side after changing available options.